TCP - reworked status codes

+++ src/protocols/TransportLayer.h
@@ -9,12 +9,12 @@ class TransportLayer {
     // this class is purely virtual and does not require explicit constructor
     
     // basic methods
-    virtual uint8_t openTransportConnection(const char* host, const char* protocol, uint16_t port, uint16_t tcpKeepAlive = 0) = 0;
-    virtual uint8_t closeTransportConnection() = 0;
-    virtual uint8_t send(const char* data) = 0;
-    virtual uint8_t send(uint8_t* data, uint32_t len) = 0;
+    virtual int16_t openTransportConnection(const char* host, const char* protocol, uint16_t port, uint16_t tcpKeepAlive = 0) = 0;
+    virtual int16_t closeTransportConnection() = 0;
+    virtual int16_t send(const char* data) = 0;
+    virtual int16_t send(uint8_t* data, uint32_t len) = 0;
     virtual size_t receive(uint8_t* data, size_t len, uint32_t timeout = 10000) = 0;
-    virtual uint16_t getNumBytes(uint32_t timeout = 10000, size_t minBytes = 10) = 0;
+    virtual size_t getNumBytes(uint32_t timeout = 10000, size_t minBytes = 10) = 0;
 };
 
 #endif
